include: move disk driver interface to own header
Move disk driver interface to own header and
separate disk access interface and disk driver interface.
