+++ drivers/disk/ramdisk.c
@@ -7,7 +7,7 @@
 #include <string.h>
 #include <zephyr/types.h>
 #include <sys/__assert.h>
-#include <disk/disk_access.h>
+#include <drivers/disk.h>
 #include <errno.h>
 #include <init.h>
 #include <device.h>
